backend/test_synthetic_faces.py
# save as: fast_test.py
import requests
import base64
import json
import io
from PIL import Image, ImageDraw
import time

API_URL = "http://localhost:5000"

def create_synthetic_face(name, color, features):
    """Create a realistic-looking synthetic face"""
    img = Image.new('RGB', (400, 400), 'beige')
    draw = ImageDraw.Draw(img)
    
    # Face oval
    draw.ellipse([100, 80, 300, 320], fill=color, outline='brown', width=3)
    
    # Hair
    if features.get('hair'):
        draw.ellipse([90, 60, 310, 180], fill='black', outline='black')
    
    # Eyes
    eye_y = 150
    draw.ellipse([140, eye_y, 170, eye_y+30], fill='white', outline='black', width=2)
    draw.ellipse([230, eye_y, 260, eye_y+30], fill='white', outline='black', width=2)
    draw.ellipse([150, eye_y+10, 160, eye_y+20], fill='black')
    draw.ellipse([240, eye_y+10, 250, eye_y+20], fill='black')
    
    # Nose
    draw.line([(200, 180), (200, 220)], fill='brown', width=3)
    draw.arc([190, 210, 210, 230], 0, 180, fill='brown', width=2)
    
    # Mouth
    if features.get('smile'):
        draw.arc([160, 240, 240, 280], 0, 180, fill='red', width=3)
    else:
        draw.line([(170, 260), (230, 260)], fill='red', width=3)
    
    # Facial hair
    if features.get('beard'):
        draw.arc([140, 260, 260, 340], 0, 180, fill='black', width=8)
    
    buffer = io.BytesIO()
    img.save(buffer, format='JPEG', quality=95)
    buffer.seek(0)
    return base64.b64encode(buffer.read()).decode()

def create_dog(breed_color):
    """Create synthetic dog"""
    img = Image.new('RGB', (400, 400), 'lightgreen')
    draw = ImageDraw.Draw(img)
    
    draw.ellipse([100, 200, 300, 350], fill=breed_color, outline='black', width=2)
    draw.ellipse([150, 120, 250, 220], fill=breed_color, outline='black', width=2)
    draw.ellipse([140, 100, 170, 140], fill=breed_color, outline='black', width=2)
    draw.ellipse([230, 100, 260, 140], fill=breed_color, outline='black', width=2)
    draw.ellipse([170, 150, 185, 165], fill='black')
    draw.ellipse([215, 150, 230, 165], fill='black')
    
    buffer = io.BytesIO()
    img.save(buffer, format='JPEG')
    buffer.seek(0)
    return base64.b64encode(buffer.read()).decode()

def create_car(color):
    """Create synthetic car"""
    img = Image.new('RGB', (400, 400), 'skyblue')
    draw = ImageDraw.Draw(img)
    
    draw.rectangle([100, 180, 300, 280], fill=color, outline='black', width=2)
    draw.rectangle([130, 130, 270, 180], fill=color, outline='black', width=2)
    draw.ellipse([110, 260, 160, 310], fill='black', outline='gray', width=3)
    draw.ellipse([240, 260, 290, 310], fill='black', outline='gray', width=3)
    
    buffer = io.BytesIO()
    img.save(buffer, format='JPEG')
    buffer.seek(0)
    return base64.b64encode(buffer.read()).decode()

print("⚡ FAST UNIVERSAL IMAGE RECOGNITION TEST")
print("=" * 70)
print("✨ No downloads required - using synthetic images")
print("=" * 70)

total_tests = 0
passed_tests = 0
total_time = 0

# ============================================================================
# PART 1: FACE TESTS
# ============================================================================
print("\n👤 PART 1: FACE RECOGNITION TESTS")
print("-" * 70)

# Test 1.1: Same person features
print("\n🔬 TEST 1: Same Person (Similar Features)")
total_tests += 1

person1_v1 = create_synthetic_face("John", 'peachpuff', {'hair': True, 'smile': True, 'beard': True})
person1_v2 = create_synthetic_face("John", 'peachpuff', {'hair': True, 'smile': False, 'beard': True})

start = time.time()
response = requests.post(f"{API_URL}/compare", json={
    "image1": person1_v1,
    "image2": person1_v2
})
elapsed = time.time() - start
total_time += elapsed

result = response.json()
similarity = result.get('similarity', 0)
match = result.get('match', False)

print(f"   Similarity: {similarity:.1f}%")
print(f"   Match: {'YES ✅' if match else 'NO ❌'}")
print(f"   Time: {elapsed:.2f}s")

if similarity >= 40:  # Realistic threshold for synthetic faces
    print(f"   ✅ PASS")
    passed_tests += 1
else:
    print(f"   ❌ FAIL")

# Test 1.2: Different people
print("\n🔬 TEST 2: Different People")
total_tests += 1

person1 = create_synthetic_face("John", 'peachpuff', {'hair': True, 'smile': True, 'beard': True})
person2 = create_synthetic_face("Jane", 'wheat', {'hair': True, 'smile': True, 'beard': False})

start = time.time()
response = requests.post(f"{API_URL}/compare", json={
    "image1": person1,
    "image2": person2
})
elapsed = time.time() - start
total_time += elapsed

result = response.json()
similarity = result.get('similarity', 0)

print(f"   Similarity: {similarity:.1f}%")
print(f"   Time: {elapsed:.2f}s")

if similarity < 90:
    print(f"   ✅ PASS - Detected difference")
    passed_tests += 1
else:
    print(f"   ⚠️  PARTIAL")

# ============================================================================
# PART 2: ANIMAL TESTS
# ============================================================================
print("\n\n🐾 PART 2: ANIMAL/PET RECOGNITION TESTS")
print("-" * 70)

# Test 2.1: Same breed
print("\n🔬 TEST 3: Same Animal Breed")
total_tests += 1

dog1 = create_dog('brown')
dog2 = create_dog('brown')

start = time.time()
response = requests.post(f"{API_URL}/compare", json={
    "image1": dog1,
    "image2": dog2
})
elapsed = time.time() - start
total_time += elapsed

result = response.json()
similarity = result.get('similarity', 0)

print(f"   Similarity: {similarity:.1f}%")
print(f"   Type: {result.get('comparison_type', 'N/A')}")
print(f"   Time: {elapsed:.2f}s")

if similarity >= 50:
    print(f"   ✅ PASS")
    passed_tests += 1
else:
    print(f"   ⚠️  PARTIAL")

# Test 2.2: Animal detection
print("\n🔬 TEST 4: Animal Detection")
total_tests += 1

dog = create_dog('brown')

start = time.time()
response = requests.post(f"{API_URL}/detect", json={"image": dog})
elapsed = time.time() - start
total_time += elapsed

result = response.json()
labels = result.get('detected', {}).get('labels', [])
labels_text = ' '.join([l['name'].lower() for l in labels[:5]])

print(f"   Primary Type: {result.get('primary_type', 'unknown')}")
print(f"   Top Labels: {labels_text[:80]}")
print(f"   Time: {elapsed:.2f}s")

if any(word in labels_text for word in ['dog', 'animal', 'pet', 'mammal']):
    print(f"   ✅ PASS - Animal detected")
    passed_tests += 1
else:
    print(f"   ⚠️  PARTIAL - Check labels above")

# ============================================================================
# PART 3: OBJECT TESTS
# ============================================================================
print("\n\n📦 PART 3: OBJECT RECOGNITION TESTS")
print("-" * 70)

# Test 3.1: Same object type
print("\n🔬 TEST 5: Same Object Type (Cars)")
total_tests += 1

car1 = create_car('red')
car2 = create_car('blue')

start = time.time()
response = requests.post(f"{API_URL}/compare", json={
    "image1": car1,
    "image2": car2
})
elapsed = time.time() - start
total_time += elapsed

result = response.json()
similarity = result.get('similarity', 0)

print(f"   Similarity: {similarity:.1f}%")
print(f"   Time: {elapsed:.2f}s")

if similarity >= 30:
    print(f"   ✅ PASS")
    passed_tests += 1
else:
    print(f"   ⚠️  PARTIAL")

# Test 3.2: Object detection
print("\n🔬 TEST 6: Object Detection")
total_tests += 1

car = create_car('red')

start = time.time()
response = requests.post(f"{API_URL}/detect", json={"image": car})
elapsed = time.time() - start
total_time += elapsed

result = response.json()
objects = result.get('detected', {}).get('objects', [])

print(f"   Primary Type: {result.get('primary_type', 'unknown')}")
print(f"   Objects Found: {len(objects)}")
print(f"   Time: {elapsed:.2f}s")

if result.get('primary_type') == 'object' or len(objects) > 0:
    print(f"   ✅ PASS")
    passed_tests += 1
else:
    print(f"   ❌ FAIL")

# ============================================================================
# FINAL RESULTS
# ============================================================================
print("\n\n" + "🏆" * 35)
print("FINAL RESULTS")
print("🏆" * 35)

accuracy = (passed_tests / total_tests * 100) if total_tests > 0 else 0
avg_time = (total_time / total_tests) if total_tests > 0 else 0

print(f"\n📊 Results:")
print(f"   ✅ Passed: {passed_tests}/{total_tests}")
print(f"   📈 Success Rate: {accuracy:.1f}%")
print(f"   ⏱️  Avg Time: {avg_time:.2f}s")

print(f"\n✅ Tested Categories:")
print(f"   👤 Faces: ✓")
print(f"   🐾 Animals: ✓")
print(f"   📦 Objects: ✓")

if accuracy >= 70:
    print(f"\n🎉 EXCELLENT - Universal API working!")
else:
    print(f"\n✅ FUNCTIONAL - API operational")

print("\n💡 This was a FAST test with synthetic images")
print("   For real-world testing, use your React Native app!")
print("=" * 70)
