backend/requirements.txt
flask
flask-cors
opencv-python
scikit-image
scikit-learn
pillow
numpy==2.1.0
tensorflow
scipy
